Fix printf alike format strings
Pointer substraction generates ptrdiff_t values, which maybe
are equal to int or maybe no. In this case we now for sure
the values will smaller than 32768, so we can safely cast it
to int.
Thanks to pancake who reported it!

+++ cc1/cpp.c
@@ -284,7 +284,7 @@ getdefs(Symbol *args[NR_MACROARG], int nargs, char *bp, size_t bufsiz)
 					break;
 			}
 			if (argp != &args[nargs]) {
-				sprintf(yytext, "@%02d@", argp - args);
+				sprintf(yytext, "@%02d@", (int) (argp - args));
 				ispar = 1;
 			}
 		}
